refactor[pkcs11]: extract signing module
This module should have things only related to signing. The main module
started to grow more to include parsing URIs, managing sessions and
objects, etc. so singing-related code was extracted.
